… pkg/spooledtempfile/spooled_bench_test.go