This pull request refactors the memory management logic in the `spooledtempfile` package to improve buffer pooling and system memory usage detection, and adds a comprehensive benchmark suite. The most significant changes include extracting and enhancing system memory usage detection **(with container/cgroup awareness)**, improving buffer pool safety and clarity, and introducing benchmarks to measure performance across various scenarios.

**Memory usage detection improvements:**

* Moved and refactored the system memory usage detection logic into a new file, `spooled_memory.go`, with support for cgroup v2, cgroup v1, and host memory detection, making the code container-aware and more robust. (`pkg/spooledtempfile/spooled_memory.go`, `pkg/spooledtempfile/spooled.go`) [[1]](diffhunk://#diff-4f5e4c23a002e9b68ffd0eb0b3bef81a5a70a771a52c2a526a78ecdf6663dfdaR1-R170) [[2]](diffhunk://#diff-a6af32bdadc35f74d527b75e16648c3933f075502488ddc0043a2dcded446cc4L312-L363)

**Buffer pool management improvements:**

* Changed the buffer pool to store pointers to byte slices, added `getPooledBuf` and `putPooledBuf` helpers for safer and clearer buffer management, and updated all usages to use these helpers instead of direct pool access. (`pkg/spooledtempfile/spooled.go`) [[1]](diffhunk://#diff-a6af32bdadc35f74d527b75e16648c3933f075502488ddc0043a2dcded446cc4L36-R55) [[2]](diffhunk://#diff-a6af32bdadc35f74d527b75e16648c3933f075502488ddc0043a2dcded446cc4L106-R120) [[3]](diffhunk://#diff-a6af32bdadc35f74d527b75e16648c3933f075502488ddc0043a2dcded446cc4L210-R224) [[4]](diffhunk://#diff-a6af32bdadc35f74d527b75e16648c3933f075502488ddc0043a2dcded446cc4L227-R249) [[5]](diffhunk://#diff-a6af32bdadc35f74d527b75e16648c3933f075502488ddc0043a2dcded446cc4L256-R267)
* Updated tests to use the new buffer pool helpers. (`pkg/spooledtempfile/spooled_test.go`) [[1]](diffhunk://#diff-4be7432a09fe448ab776901af2e2896827473d73bf713efc5a5a1658b1850e9dL506-R506) [[2]](diffhunk://#diff-4be7432a09fe448ab776901af2e2896827473d73bf713efc5a5a1658b1850e9dL562-R562)

**Benchmarking:**

* Added a new benchmark suite to `spooled_bench_test.go` covering in-memory, spill-to-disk, full-on-disk, high-memory-signal, parallel, and hash-read scenarios, providing extensive performance coverage. (`pkg/spooledtempfile/spooled_bench_test.go`)

These changes make the package more robust in containerized environments, improve buffer reuse safety (fixes SA6002 warnings), and provide tools for performance analysis.

Before pool buffer changes:
```
goos: darwin
goarch: arm64
pkg: github.com/internetarchive/gowarc/pkg/spooledtempfile
cpu: Apple M4
BenchmarkWrite_InMemory_Small-10                   10932            108007 ns/op        2427.10 MB/s     1095746 B/op        114 allocs/op
BenchmarkWrite_SpillToDisk-10                        580           1802795 ns/op        2326.56 MB/s     4786692 B/op        146 allocs/op
BenchmarkWrite_FullOnDisk-10                         831           1388372 ns/op        3021.02 MB/s        1225 B/op         23 allocs/op
BenchmarkWrite_SpillDueToHighMemSignal-10           5242            301173 ns/op         870.41 MB/s         585 B/op         10 allocs/op
BenchmarkRead_AfterWrite_InMemory-10              157347              7584 ns/op        34567.33 MB/s          0 B/op          0 allocs/op
BenchmarkRead_AfterWrite_OnDisk-10                  5414            208808 ns/op        40173.76 MB/s          0 B/op          0 allocs/op
BenchmarkReadAt_OnDisk-10                         605002              1954 ns/op        33534.04 MB/s          0 B/op          0 allocs/op
BenchmarkParallel_CreateAndWrite_512KB-10           3662            299933 ns/op        1748.02 MB/s     4421662 B/op        235 allocs/op
BenchmarkRead_Hash_OnDisk-10                          86          11994375 ns/op        2797.51 MB/s         160 B/op          2 allocs/op
BenchmarkCreateClose_NoWrite-10                 32111533                35.94 ns/op          120 B/op          2 allocs/op
PASS
ok      github.com/internetarchive/gowarc/pkg/spooledtempfile   13.083s
```

After pool buffer changes:

```
goos: darwin
goarch: arm64
pkg: github.com/internetarchive/gowarc/pkg/spooledtempfile
cpu: Apple M4
BenchmarkWrite_InMemory_Small-10                   10814            108188 ns/op        2423.05 MB/s     1095474 B/op        114 allocs/op
BenchmarkWrite_SpillToDisk-10                        885           1414374 ns/op        2965.48 MB/s     4786899 B/op        146 allocs/op
BenchmarkWrite_FullOnDisk-10                         963           1253683 ns/op        3345.59 MB/s        1283 B/op         23 allocs/op
BenchmarkWrite_SpillDueToHighMemSignal-10           6628            184917 ns/op        1417.63 MB/s         604 B/op         10 allocs/op
BenchmarkRead_AfterWrite_InMemory-10              154537              7254 ns/op        36138.70 MB/s          0 B/op          0 allocs/op
BenchmarkRead_AfterWrite_OnDisk-10                  5253            213364 ns/op        39315.95 MB/s          0 B/op          0 allocs/op
BenchmarkReadAt_OnDisk-10                         473926              2205 ns/op        29715.52 MB/s          0 B/op          0 allocs/op
BenchmarkParallel_CreateAndWrite_512KB-10           3932            302094 ns/op        1735.51 MB/s     4421353 B/op        235 allocs/op
BenchmarkRead_Hash_OnDisk-10                          86          12157159 ns/op        2760.06 MB/s         160 B/op          2 allocs/op
BenchmarkCreateClose_NoWrite-10                 32153550                36.02 ns/op          120 B/op          2 allocs/op
PASS
ok      github.com/internetarchive/gowarc/pkg/spooledtempfile   12.794s
```